# API/Rendu API/Nouveau dossier/API/openapi-main.yaml
openapi: 3.0.0
info:
  title: API Documentation
  version: 1.0.0
  description: This is the API documentation for the application.
servers:
  - url: http://localhost:3001
    description: Local server
paths:
...
  /category:
    get:
      tags:
        - Category
          - User
      operationId: listCategories
      summary: "List all categories"
      parameters:
        - name: "sortBy"
          in: "query"
          description: "Sort by field"
          required: false
          schema:
            type: "string"
        - name: "sortDirection"
          in: "query"
          description: "Sort direction"
          required: false
          schema:
            type: "string"
            enum: ["asc", "desc"]
      responses:
        '200':
          description: A list of categories
    post:
      tags:
        - Category
          - Admin
      operationId: createCategory
      summary: "Create a new category"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, description]
              properties:
                name:
                  type: string
                description:
                  type: string
      responses:
        '200':
          description: Category created
  /category/{id}:
    get:
      tags:
        - Category
          - User
      operationId: getCategoryById
      summary: "Get category by ID"
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Category found
        '404':
          description: Category not found
    delete:
      tags:
        - Category
          - Admin
      operationId: deleteCategory
      summary: "Delete category by ID"
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Category deleted
        '404':
          description: Category not found
    put:
      tags:
        - Category
          - Admin
      operationId: updateCategory
      summary: "Update category by ID"
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                description:
                  type: string
      responses:
        '200':
          description: Category updated successfully
        '400':
          description: Invalid category data
        '404':
          description: Category not found
  /category/name/{name}:
    get:
      tags:
        - Category
          - User
      operationId: getCategoryByName
      summary: "Get category by name"
      parameters:
        - name: name
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Category found
        '404':
          description: Category not found
...
  /bill:
    get:
      tags:
        - Bill
          - Admin
      operationId: listBills
      summary: "List all bills"
      parameters:
        - name: "sortBy"
          in: "query"
          description: "Sort by field"
          required: false
          schema:
            type: "string"
        - name: "sortDirection"
          in: "query"
          description: "Sort direction"
          required: false
          schema:
            type: "string"
            enum: ["asc", "desc"]
      responses:
        '200':
          description: A list of bills
    post:
      tags:
        - Bill
          - Admin
      operationId: createBill
      summary: "Create a new bill"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [id_user, id_reason, amount, payment_date, status]
              properties:
                id_user:
                  type: integer
                id_reason:
                  type: integer
                amount:
                  type: number
                payment_date:
                  type: string
                status:
                  type: string
      responses:
        '200':
          description: Bill created
        '400':
          description: Error creating bill
  /bill/{id}:
    get:
      tags:
        - Bill
          - Admin
      operationId: getBillById
      summary: "Get bill by ID"
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Bill found
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                  amount:
                    type: number
                  payment_date:
                    type: string
                  status:
                    type: string
                  user:
                    type: object
                    properties:
                      id:
                        type: integer
                      name:
                        type: string
                      email:
                        type: string
                  reason:
                    type: object
                    properties:
                      id:
                        type: integer
                      label:
                        type: string
                      description:
                        type: string
        '404':
          description: Bill not found
        '400':
          description: Error retrieving bill
    delete:
      tags:
        - Bill
          - Admin
      operationId: deleteBillById
      summary: "Delete bill by ID"
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Bill deleted
        '404':
          description: Bill not found
        '400':
          description: Error deleting bill
    put:
      tags:
        - Bill
          - Admin
      operationId: updateBill
      summary: "Update bill by ID"
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                id_user:
                  type: integer
                id_reason:
                  type: integer
                amount:
                  type: number
                payment_date:
                  type: string
                status:
                  type: string
      responses:
        '200':
          description: Bill updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                  amount:
                    type: number
                  payment_date:
                    type: string
                  status:
                    type: string
                  user:
                    type: object
                    properties:
                      id:
                        type: integer
                      name:
                        type: string
                      email:
                        type: string
                  reason:
                    type: object
                    properties:
                      id:
                        type: integer
                      label:
                        type: string
                      description:
                        type: string
        '400':
          description: Invalid bill data
        '404':
          description: Bill not found
  /bill/user/{userId}:
    get:
      tags:
        - Bill
          - Admin
      operationId: getBillByUserId
      summary: "Get bills by user ID"
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Bills found
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: integer
                    amount:
                      type: number
                    payment_date:
                      type: string
                    status:
                      type: string
                    user:
                      type: object
                      properties:
                        id:
                          type: integer
                        name:
                          type: string
                        email:
                          type: string
                    reason:
                      type: object
                      properties:
                        id:
                          type: integer
                        label:
                          type: string
                        description:
                          type: string
        '404':
          description: No bills found for the given user
        '400':
          description: Error retrieving bills
  /bill/reason/{reasonId}:
    get:
      tags:
        - Bill
          - Admin
      operationId: getBillByReasonId
      summary: "Get bills by reason ID"
      parameters:
        - name: reasonId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Bills found
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: integer
                    amount:
                      type: number
                    payment_date:
                      type: string
                    status:
                      type: string
                    user:
                      type: object
                      properties:
                        id:
                          type: integer
                        name:
                          type: string
                        email:
                          type: string
                    reason:
                      type: object
                      properties:
                        id:
                          type: integer
                        label:
                          type: string
                        description:
                          type: string
        '404':
          description: No bills found for the given reason
        '400':
          description: Error retrieving bills
  /bill/status/{status}:
    get:
      tags:
        - Bill
          - Admin
      operationId: getBillByStatus
      summary: "Get bills by status"
      parameters:
        - name: status
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Bills found
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: integer
                    amount:
                      type: number
                    payment_date:
                      type: string
                    status:
                      type: string
                    user:
                      type: object
                      properties:
                        id:
                          type: integer
                        name:
                          type: string
                        email:
                          type: string
                    reason:
                      type: object
                      properties:
                        id:
                          type: integer
                        label:
                          type: string
                        description:
                          type: string
        '404':
          description: No bills found for the given status
        '400':
          description: Error retrieving bills
...